+++ mapbox_baselayer/views.py
@@ -10,14 +10,14 @@ from mapbox_baselayer.utils import DEFAULT_OSM_TILEJSON, get_map_base_layers
 
 
 class MapboxBaseLayerJsonDetailView(BaseDetailView):
-    queryset = models.MapBaseLayer.objects.exclude(
-        base_layer_type="mapbox"
-    )  # mapbox provide its own json
+    queryset = models.MapBaseLayer.objects.all()
 
     def get(self, request, *args, **kwargs):
         tilejson = deepcopy(self.get_object().tilejson)
-        glyphs_url = request.build_absolute_uri(tilejson["glyphs"])
-        tilejson["glyphs"] = unquote(glyphs_url)
+        glyphs_url = tilejson.get("glyphs")
+        if glyphs_url and not glyphs_url.startswith(("http", "mapbox")):
+            glyphs_url = request.build_absolute_uri(tilejson["glyphs"])
+            tilejson["glyphs"] = unquote(glyphs_url)
         return JsonResponse(tilejson)